pyproject.toml
[build-system]
requires = ['setuptools>=45', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = "priorpatch"
version = "0.1.0"
description = "Math-only forensic toolkit for detecting AI-generated or manipulated images"
readme = "README.md"
authors = [
    {name = "Jayashankar VR", email = "jayashankarvr@users.noreply.github.com"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Security",
    "Typing :: Typed",
]
keywords = ["forensics", "image-analysis", "deepfake-detection", "signal-processing"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0,<2.0.0",
    "Pillow>=9.0.0,<11.0.0",
    "matplotlib>=3.5.0,<4.0.0",
    "scipy>=1.7.0,<2.0.0",
    "tqdm>=4.65.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mkdocs>=1.4.0",
    "mkdocs-material>=9.0.0",
]

[project.urls]
Homepage = "https://github.com/jayashankarvr/priorpatch"
Documentation = "https://github.com/jayashankarvr/priorpatch#readme"
Repository = "https://github.com/jayashankarvr/priorpatch"
Issues = "https://github.com/jayashankarvr/priorpatch/issues"
Changelog = "https://github.com/jayashankarvr/priorpatch/blob/main/CHANGELOG.md"

[project.scripts]
priorpatch = "priorpatch.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=priorpatch --cov-report=term-missing"

[tool.coverage.run]
source = ["priorpatch"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["priorpatch*"]
exclude = ["tests*", "config*", "results*", "docs*", "examples*", "benchmarks*"]
